pyproject.toml
[project]
name = "tochka-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "everbase",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "orjson>=3.10.18",
    "pydantic-settings>=2.9.1",
]

[tool.uv.sources]
everbase = { git = "https://github.com/OneTells/everbase.git" }
